Refactor - unify Home screen implementation
