+++ frontend/src/screens/Home/StakePoolsInfo.js
@@ -2,19 +2,13 @@
 
 import * as React from 'react'
 import gql from 'graphql-tag'
-import {Grid} from '@material-ui/core'
 import {defineMessages} from 'react-intl'
 import {useQuery} from 'react-apollo-hooks'
 
-import {
-  KeyValueCard,
-  SimpleLayout,
-  LoadingOverlay,
-  ErrorOverlay,
-  Overlay,
-  AdaValue,
-  LoadingError,
-} from '@/components/visual'
+import InfoCard from './InfoCard'
+import Layout from './Layout'
+
+import {SimpleLayout, AdaValue, LoadingError} from '@/components/visual'
 import {useI18n} from '@/i18n/helpers'
 import {useCurrentEpoch} from './common'
 
@@ -49,11 +43,6 @@ const cardHeaders = defineMessages({
   usedInEpoch: 'Used in epoch {epoch}',
   willBeUsedInEpoch: 'Will be used in epoch {epoch}',
 })
-
-type Field = {|
-  label: string,
-  getValue: Function,
-|}
 
 const getAllTimeFields = ({translate, formatInt}) => [
   {
@@ -92,41 +81,6 @@ const getPerEpochFields = ({translate, formatInt}) => [
     getValue: (data) => <AdaValue value={data.rewardsPerBlock} />,
   },
 ]
-
-const Row = ({children}: {children: React.Node}) => (
-  <Grid container direction="row" justify="space-around" spacing={24}>
-    {children}
-  </Grid>
-)
-
-type InfoCardProps = {|
-  data: Object, // TODO: get graphql type
-  fields: Array<Field>,
-  cardLabel: string,
-  cardValue: string,
-  loading: boolean,
-  error: any,
-|}
-
-const InfoCard = ({data, fields, cardLabel, cardValue, loading, error}: InfoCardProps) => {
-  const items = fields.map((fieldDesc) => ({
-    label: fieldDesc.label,
-    value: data && fieldDesc.getValue(data),
-  }))
-
-  return (
-    <Overlay.Wrapper className="h-100">
-      <KeyValueCard
-        className="h-100"
-        header={<KeyValueCard.Header label={cardLabel} value={cardValue} />}
-      >
-        <KeyValueCard.Body items={items} />
-        <LoadingOverlay loading={loading} />
-        <ErrorOverlay error={!loading && error} />
-      </KeyValueCard>
-    </Overlay.Wrapper>
-  )
-}
 
 const AllTimeCard = () => {
   const formatters = useI18n()
@@ -204,13 +158,14 @@ export default () => {
     )
   }
 
+  const {Section, Row, Item} = Layout
   return (
-    <SimpleLayout title={tr(messages.header)} maxWidth="1200px">
+    <Section title={tr(messages.header)}>
       <Row>
-        <Grid item xs={6}>
+        <Item>
           <AllTimeCard />
-        </Grid>
-        <Grid item xs={6}>
+        </Item>
+        <Item>
           {/* For now not solving case when there is no `previousEpoch`. Possibly
               we will show only `N/A` values. Depends on server implementation. */}
           <PerEpochCard
@@ -225,10 +180,10 @@ export default () => {
             cardLabel={tr(cardHeaders.previousSnapshot)}
             epoch={currentEpoch !== null ? currentEpoch - 1 : null}
           />
-        </Grid>
+        </Item>
       </Row>
       <Row>
-        <Grid item xs={6}>
+        <Item>
           <PerEpochCard
             cardValue={tr(cardHeaders.usedInEpoch, {
               epoch: currentEpoch != null ? currentEpoch : '',
@@ -236,8 +191,8 @@ export default () => {
             cardLabel={tr(cardHeaders.currentSnapshot)}
             epoch={currentEpoch != null ? currentEpoch : null}
           />
-        </Grid>
-        <Grid item xs={6}>
+        </Item>
+        <Item>
           <PerEpochCard
             cardValue={tr(cardHeaders.willBeUsedInEpoch, {
               epoch: currentEpoch != null ? currentEpoch + 1 : '',
@@ -245,8 +200,8 @@ export default () => {
             cardLabel={tr(cardHeaders.nextSnapshot)}
             epoch={currentEpoch !== null ? currentEpoch + 1 : null}
           />
-        </Grid>
+        </Item>
       </Row>
-    </SimpleLayout>
+    </Section>
   )
 }
